feat(data-store): initialize DB using migrations (#679)
* feat(data-store): initialize DB using migrations
* fix(data-store): align onDelete cascade settings and do some decoupling of tests
* test(data-store): use multiple DB connection options in tests
* test(ci): use postgres in CI tests

fixes #676

BREAKING CHANGE: database needs migrations for initialization. See #679 #676
The `@veramo/data-store` package relies on `typeorm` as a database abstraction.
Typeorm has a connection flag `synchonize` which bootstraps the database along with schema and relations based on a set of `Entities` (annotated typescript classes).
This is very handy for fast development iterations but it is **not recommended for production** use because there is too much ambiguity possible when the `Entities` change, and there is a risk of data loss.
The recommended way to do things is to use the `migrations` mechanism. It allows you to migrate to new database schemas when necessary, and even customize the database to your own needs.

**Going forward, this is the mechanism we will be recommending for connections.**